backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==8.3.2
httpx==0.27.0
transformers==4.41.2
torch==2.3.1
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2
numpy==1.26.4
python-multipart==0.0.9
prefect==3.0.2
